flask-app/requirements.txt
Flask==3.1.0
psycopg2-binary==2.9.10
kubernetes==32.0.1